requirements.txt
pip
opencv_python_headless
streamlit==1.35.0
numpy
tensorflow
Pillow
